Move _utils content
