+++ transactron/utils/__init__.py
@@ -1,3 +1,5 @@
 from .utils import *  # noqa: F401
 from ._typing import *  # noqa: F401
 from .debug_signals import *  # noqa: F401
+from .modules import *  # noqa: F401
+from .assign import *  # noqa: F401
